Move transformJsx to own file
